requirements.txt
# BP010 Auditoría - dependencias con rangos compatibles (recomendado: usar venv)
pandas>=2.0,<3
sqlalchemy>=2.0,<3
psycopg2-binary>=2.9
ipykernel>=6
nbformat>=5
nbconvert>=7
python-dotenv>=1.0
requests>=2.28
openpyxl>=3.0
